Added single char token parsing

+++ src/error_handling.rs
@@ -7,12 +7,12 @@ impl ErrorReporter {
         Self { had_error: false }
     }
 
-    pub fn report(&mut self, line: u32, location: &str, message: &str) {
+    pub fn report(&mut self, line: usize, location: &str, message: &str) {
         println!("[line {line}] Error {location}: {message}");
         self.had_error = true;
     }
 
-    pub fn error(&mut self, line: u32, message: &str) {
+    pub fn error(&mut self, line: usize, message: &str) {
         self.report(line, "", message);
     }
 }
